% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_stat.R
\name{cn_stat}
\alias{cn_stat}
\title{Intergrate alpha hat and beta hat into Z-statistics to apply the local approaches later}
\usage{
cn_stat(PS, trans = T, mc = 5)
}
\arguments{
\item{PS}{The pre-fitting results (optional when using VCT, use for determining the signs of the statistics if available)}

\item{trans}{An indicator to apply decorrelation or not, default=TRUE.}

\item{mc}{Number of cores for parallel computing, default=5.}
}
\description{
Intergrate alpha hat and beta hat into Z-statistics to apply the local approaches later
}
\examples{
HA=sim_mediation_data(hypo="HA",mm=0.1,vv=0.1,sm=10)
PS=pre_stat(HA$S,HA$M,HA$Y,HA$X)
CSp=cn_stat(PS,trans = T,mc)
CSm=cn_stat(PS,trans = F,mc)
}
