% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCN.R
\name{GCN}
\alias{GCN}
\title{Global approach: TSQ, GBJ, GHC, minP}
\usage{
GCN(PS, method = "minP", mc = 5, prec = 0.001)
}
\arguments{
\item{PS}{The pre-fitting results}

\item{method}{Other approcahes beside VCT, here we provide "TSQ","GBJ","GHC", and "minP", default="minP".}

\item{mc}{Number of cores for parallel computing, default=5.}

\item{prec}{The unit to construct empirical normal product pdf using in composite test, default=1e-3.}
}
\description{
Global approach: TSQ, GBJ, GHC, minP
}
\examples{
HA=sim_mediation_data(hypo="HA",mm=0.1,vv=0.1,sm=10)
PS=pre_stat(HA$S,HA$M,HA$Y,HA$X)
TSQg=GCN(PS,"TSQ")
GBJg=GCN(PS,"GBJ")
GHCg=GCN(PS,"GHC")
mnPg=GCN(PS,"minP")
}
