% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCN_VCT.R
\name{GCN_VCT}
\alias{GCN_VCT}
\title{Global approach: VCT}
\usage{
GCN_VCT(S, M, Y, X, PS = NULL, mc = 5, prec = 0.001)
}
\arguments{
\item{S, M, Y, X}{Input data: exposure(vector), mediators(list of vectors), outcomes(list), and covariates(vectors).}

\item{PS}{The pre-fitting results (optional when using VCT, use for determining the signs of the statistics if available)}

\item{mc}{Number of cores for parallel computing, default=5.}

\item{prec}{The unit to construct empirical normal product pdf using in composite test, default=1e-3.}
}
\description{
Global approach: VCT
}
\examples{
HA=sim_mediation_data(hypo="HA",mm=0.1,vv=0.1,sm=10)
PS=pre_stat(HA$S,HA$M,HA$Y,HA$X)
VCTg=GCN_VCT(HA$S,HA$M,HA$Y,HA$X,PS)
}
