% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCN.R
\name{LCN}
\alias{LCN}
\title{Local approach: BJ, HC, minP}
\usage{
LCN(CS, method = "minP", mc = 5)
}
\arguments{
\item{method}{We provide "BJ","HC", and "minP", default="minP".}

\item{mc}{Number of cores for parallel computing, default=5.}

\item{PS}{The pre-fitting results (optional when using VCT, use for determining the signs of the statistics if available)}
}
\description{
Local approach: BJ, HC, minP
}
\examples{
HA=sim_mediation_data(hypo="HA",mm=0.1,vv=0.1,sm=10)
PS=pre_stat(HA$S,HA$M,HA$Y,HA$X)
CSp=cn_stat(PS,trans = T)
BJp=LCN(CSp,"BJ")
HCp=LCN(CSp,"HC")
MPp=LCN(CSp,"minP")
CSm=cn_stat(PS,trans = F)
BJm=LCN(CSm,"BJ")
HCm=LCN(CSm,"HC")
MPm=LCN(CSm,"minP")
}
