% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_stat.R
\name{pre_stat}
\alias{pre_stat}
\title{Pre-fitting the data, produce alpha hat and beta hat, required for every tests excepts VCT.}
\usage{
pre_stat(S, M, Y, X, mc = 5)
}
\arguments{
\item{S, M, Y, X}{Input data: exposure(vector), mediators(list of vectors), outcomes(list), and covariates(vectors).}

\item{mc}{Number of cores for parallel computing, default=5.}
}
\description{
Pre-fitting the data, produce alpha hat and beta hat, required for every tests excepts VCT.
}
\examples{
HA=sim_mediation_data(hypo="HA",mm=0.1,vv=0.1,sm=10)
PS=pre_stat(HA$S,HA$M,HA$Y,HA$X)
}
