% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mediation_data.R
\name{sim_mediation_data}
\alias{sim_mediation_data}
\title{Simulate data for analysis.}
\usage{
sim_mediation_data(
  hypo = "H00",
  sample_size = 1000,
  num_mediators = 30,
  num_mechanisms = 10000,
  num_covariates = 2,
  SS = NULL,
  har = 0.1,
  mm = 0,
  vv = 0,
  sm = 5,
  mc = 5,
  Styp = "C"
)
}
\arguments{
\item{hypo}{Hypothesis type: "H00" refers to the complete null hypothese (default);
"G0A" refers to the null hypothese with only alpha-signals;
"G0B" refers to the null hypothese with only beta-signals;
"H02" refers to the cases with disjoint signals, if rho=0 then it is a null, otherwise alternative;
"HA" refer to the alternative hypothesis with connected signals.}

\item{sample_size}{Number of samples, default=1000.}

\item{num_mediators}{Number of mediators, default=30.}

\item{num_mechanisms}{Number of mechanisms, default=10000.}

\item{num_covariates}{Number of covariates, default=2.}

\item{SS}{The correlation matrix between mediators.}

\item{har}{The percentage of mechanisms with signals within one experiment, default=0.1. (affect when hypo is not "H00")}

\item{mm}{The mean of the signal, default=0. (affect when hypo is not "H00")}

\item{vv}{The variance of the signal, default=0. (affect when hypo is not "H00")}

\item{sm}{Number of mediators with signal within one mechanism, default=5. (affect when hypo is not "H00")}

\item{mc}{Number of cores for parallel computing, default=5.}
}
\description{
Simulate data for analysis.
}
\examples{
H00=sim_mediation_data()
H00_ss=sim_mediation_data(sample_size=100)
H00_rh=sim_mediation_data(rho=0)
G0A=sim_mediation_data(hypo="G0A",har=0.1,mm=0,vv=0.05,sm=5)
G0B=sim_mediation_data(hypo="G0B",har=0.1,mm=0,vv=0.05,sm=5)
H02_null=sim_mediation_data(hypo="H02",har=1,mm=0,vv=0.05,sm=2)
HA=sim_mediation_data(hypo="HA",mm=0.1,vv=0.1,sm=10)
}
